Refactor to MatchService
